feat: add to check if i am logged in or not before we start packaging project

+++ src/commands/deploy.js
@@ -1,31 +1,39 @@
 import fs from 'fs';
 import ora from 'ora';
+import * as config from '../utils/config.js';
 import * as logger from '../utils/logger.js';
 import { zipDirectory } from '../lib/zipper.js';
 import { deploy } from '../lib/api.js';
 
 export default async function deployCommand() {
-    const spinner = ora('Packaging your project...').start();
+    const token = config.get('auth.token');
+    
+    if(!token){
+        logger.error('login first');
+    }
+    else{
+        const spinner = ora('Packaging your project...').start();
 
-    let zipPath = null;
+        let zipPath = null;
 
-    try {
+        try {
 
-        zipPath = await zipDirectory();
-        spinner.succeed('Project packed!');
+            zipPath = await zipDirectory();
+            spinner.succeed('Project packed!');
 
-        spinner.start('Uploading to Dock Hosting...');
-        const result = await deploy(zipPath);
+            spinner.start('Uploading to Dock Hosting...');
+            const result = await deploy(zipPath);
 
-        spinner.succeed('Deployment complete!');
-        logger.success(`Your site is live at: ${result.url}`);
+            spinner.succeed('Deployment complete!');
+            logger.success(`Your site is live at: ${result.url}`);
 
-    } catch (error) {
-        spinner.fail('Deployment failed.');
-        logger.error(error.message);
-    } finally {
-        if (zipPath && fs.existsSync(zipPath)) {
-            fs.unlinkSync(zipPath);
-        }
+        } catch (error) {
+            spinner.fail('Deployment failed.');
+            logger.error(error.message);
+        } finally {
+            if (zipPath && fs.existsSync(zipPath)) {
+                fs.unlinkSync(zipPath);
+            }
+        }     
     }
 }
